doc: Add TODO as a reminder for important features
 - Integrate contact handling (get-set contact and use
it for SMS and phone calls): SMS should return the
contact name too and sending an SMS should be possible
with the contact name. Getting a list of all contacts
stored in the GSM module should also be possible.

 - Add python demo forwarding all incoming SMS to specific
email or phone number: This should be done after the contact
handling