lib: Send multiple SMS if Text Mode is used and SMS is too big
This feature will only send multiple SMS and not send
"Multipart Message" (MMS). It will allow to send multiple SMS
instead of sending it at all. In text mode, it is not possible to
do better than that.

This is not #15 feature since it doesn't handle multipart message and
PDU mode multipart message is not handled yet.